requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# embedding local model
sentence-transformers

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
torch
streamlit
langchain_community
pypdf
bs4

# Vector Store Chroma
chromadb
tiktoken

# Retrievers
wikipedia
faiss-

# Project
youtube-transcript-api

# Tools
duckduckgo-search
langchain_experimental

# Tool Calling
requests